+++ CHANGELOG.md
@@ -11,6 +11,8 @@ to a section with the version name.
 * Update version (`-v`) to expose exact version in the payload instead of just a display string. This allows clients to
   parse and check it to decide whether certain features are supported
 * Improve error message for unknown commands
+* Enhance validate command to expose more details
+  * We now expose structured dictionaries of (code, error) and (code, warning) for both errors and warnings
 
 ## 2.0.1